TweetStreamAnalyseLambda/requirements.txt
asn1crypto==0.23.0
boto3==1.4.7
botocore==1.7.19
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
cryptography==2.1.3
docutils==0.14
idna==2.6
jmespath==0.9.3
kafka-python==1.3.5
pip==9.0.1
pycparser==2.18
pyOpenSSL==17.3.0
pysolr==3.6.0
python-dateutil==2.6.1
requests==2.18.4
s3transfer==0.1.11
setuptools==38.2.1
six==1.11.0
twitter==1.17.1
urllib3==1.22
watson-developer-cloud==1.0.0
